requirements.txt
PyYAML
elasticsearch
certifi
faker
simplejson
pytz